Add and update Javadoc
