+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/ArgumentCountValidationMode.java
@@ -14,39 +14,47 @@ import org.apiguardian.api.API;
 import org.junit.jupiter.params.provider.ArgumentsSource;
 
 /**
- * Enumeration of argument count validation modes for {@link ParameterizedTest @ParameterizedTest}.
+ * Enumeration of argument count validation modes for
+ * {@link ParameterizedContainer @ParameterizedContainer} and
+ * {@link ParameterizedTest @ParameterizedTest}.
  *
- * <p>When an {@link ArgumentsSource} provides more arguments than declared by the test method,
- * there might be a bug in the test method or the {@link ArgumentsSource}.
- * By default, the additional arguments are ignored.
- * {@link ArgumentCountValidationMode} allows you to control how additional arguments are handled.
+ * <p>When an {@link ArgumentsSource} provides more arguments than declared by
+ * the parameterized container or method, there might be a bug in the
+ * class/method or the {@link ArgumentsSource}. By default, the additional
+ * arguments are ignored. {@link ArgumentCountValidationMode} allows you to
+ * control how additional arguments are handled.
  *
  * @since 5.12
- * @see ParameterizedTest
+ * @see ParameterizedContainer#argumentCountValidation()
+ * @see ParameterizedTest#argumentCountValidation()
  */
 @API(status = API.Status.EXPERIMENTAL, since = "5.12")
 public enum ArgumentCountValidationMode {
+
 	/**
 	 * Use the default validation mode.
 	 *
 	 * <p>The default validation mode may be changed via the
-	 * {@value ArgumentCountValidator#ARGUMENT_COUNT_VALIDATION_KEY} configuration parameter
-	 * (see the User Guide for details on configuration parameters).
+	 * {@value ArgumentCountValidator#ARGUMENT_COUNT_VALIDATION_KEY}
+	 * configuration parameter (see the User Guide for details on configuration
+	 * parameters).
 	 */
 	DEFAULT,
 
 	/**
 	 * Use the "none" argument count validation mode.
 	 *
-	 * <p>When there are more arguments provided than declared by the test method,
-	 * these additional arguments are ignored.
+	 * <p>When there are more arguments provided than declared by the
+	 * parameterized container or method, these additional arguments are
+	 * ignored.
 	 */
 	NONE,
 
 	/**
 	 * Use the strict argument count validation mode.
 	 *
-	 * <p>When there are more arguments provided than declared by the test method, this raises an error.
+	 * <p>When there are more arguments provided than declared by the
+	 * parameterized container or method, this raises an error.
 	 */
 	STRICT,
 }
