+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/FieldParameterDeclaration.java
@@ -13,8 +13,8 @@ package org.junit.jupiter.params;
 import java.lang.reflect.Field;
 import java.util.Optional;
 
-import org.junit.jupiter.params.provider.ParameterDeclaration;
 import org.junit.jupiter.params.support.FieldContext;
+import org.junit.jupiter.params.support.ParameterDeclaration;
 
 /**
  * @since 5.13
