+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/Parameter.java
@@ -21,13 +21,40 @@ import java.lang.annotation.Target;
 import org.apiguardian.api.API;
 
 /**
+ * {@code @Parameter} is used to signal that a field in a
+ * <em>parameterized container</em> class constitutes a <em>parameter</em> and
+ * marks it for field injection.
+ *
+ * <p>{@code @Parameter} may also be used as a meta-annotation in order to
+ * create a custom <em>composed annotation</em> that inherits the semantics of
+ * {@code @Parameter}.
+ *
  * @since 5.13
  * @see ParameterizedContainer
+ * @see org.junit.jupiter.params.aggregator.ArgumentsAccessor
+ * @see org.junit.jupiter.params.aggregator.AggregateWith
+ * @see org.junit.jupiter.params.converter.ArgumentConverter
+ * @see org.junit.jupiter.params.converter.ConvertWith
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.ANNOTATION_TYPE, ElementType.FIELD })
 @Documented
 @API(status = EXPERIMENTAL, since = "5.13")
 public @interface Parameter {
+
+	/**
+	 * {@return the index of the parameter in the list of parameters}
+	 *
+	 * <p>Must be {@code -1} (the default) for <em>aggregators</em>,
+	 * that is any field of type
+	 * {@link org.junit.jupiter.params.aggregator.ArgumentsAccessor ArgumentsAccessor}
+	 * or any field annotated with
+	 * {@link org.junit.jupiter.params.aggregator.AggregateWith @AggregateWith}.
+	 *
+	 * <p>May be omitted if there's a single <em>indexed parameter</em>.
+	 * Otherwise, must be unique among all <em>indexed parameters</em> of the
+	 * parameterized container class and its superclasses.
+	 */
 	int value() default -1;
+
 }
