+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/aggregator/AggregateWith.java
@@ -24,10 +24,14 @@ import org.apiguardian.api.API;
  * {@code @AggregateWith} is an annotation that allows one to specify an
  * {@link ArgumentsAggregator}.
  *
- * <p>This annotation may be applied to a parameter of a
+ * <p>This annotation may be applied to parameters of a
+ * {@link org.junit.jupiter.params.ParameterizedContainer @ParameterizedContainer}
+ * constructor or its
+ * {@link org.junit.jupiter.params.Parameter @Parameter}-annotated fields, or to
+ * parameters of a
  * {@link org.junit.jupiter.params.ParameterizedTest @ParameterizedTest} method
  * in order for an aggregated value to be resolved for the annotated parameter
- * when the test method is invoked.
+ * when the parameterized container or method is invoked.
  *
  * <p>{@code @AggregateWith} may also be used as a meta-annotation in order to
  * create a custom <em>composed annotation</em> that inherits the semantics
