+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/converter/ConvertWith.java
@@ -23,9 +23,13 @@ import org.apiguardian.api.API;
 /**
  * {@code @ConvertWith} is an annotation that allows one to specify an explicit
  * {@link ArgumentConverter}.
-
- * <p>This annotation may be applied to parameters of
- * {@link org.junit.jupiter.params.ParameterizedTest @ParameterizedTest} methods
+ *
+ * <p>This annotation may be applied to parameters of a
+ * {@link org.junit.jupiter.params.ParameterizedContainer @ParameterizedContainer}
+ * constructor or its
+ * {@link org.junit.jupiter.params.Parameter @Parameter}-annotated fields, or to
+ * parameters of a
+ * {@link org.junit.jupiter.params.ParameterizedTest @ParameterizedTest} method
  * which need to have their {@code Arguments} converted before consuming them.
  *
  * @since 5.0
