+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/provider/MethodSource.java
@@ -20,6 +20,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 import org.apiguardian.api.API;
+import org.junit.jupiter.params.ParameterizedContainer;
 import org.junit.jupiter.params.ParameterizedTest;
 
 /**
@@ -32,6 +33,7 @@ import org.junit.jupiter.params.ParameterizedTest;
  * <p>Each factory method must generate a <em>stream</em> of <em>arguments</em>,
  * and each set of "arguments" within the "stream" will be provided as the physical
  * arguments for individual invocations of the annotated
+ * {@code @ParameterizedContainer} class or
  * {@link ParameterizedTest @ParameterizedTest} method. Generally speaking this
  * translates to a {@link java.util.stream.Stream Stream} of {@link Arguments}
  * (i.e., {@code Stream<Arguments>}); however, the actual concrete return type
@@ -92,6 +94,10 @@ import org.junit.jupiter.params.ParameterizedTest;
  * test instance lifecycle mode is used; whereas, factory methods in external
  * classes must always be {@code static}.
  *
+ * <p>This behavior and the above examples also apply to parameters of
+ * {@link ParameterizedContainer @ParameterizedContainer} classes, regardless
+ * whether field or constructor injection is used.
+ *
  * <p>Factory methods can declare parameters, which will be provided by registered
  * implementations of {@link org.junit.jupiter.api.extension.ParameterResolver}.
  *
@@ -99,6 +105,7 @@ import org.junit.jupiter.params.ParameterizedTest;
  * @see FieldSource
  * @see Arguments
  * @see ArgumentsSource
+ * @see ParameterizedContainer
  * @see ParameterizedTest
  * @see org.junit.jupiter.api.TestInstance
  */
@@ -131,7 +138,9 @@ public @interface MethodSource {
 	 *
 	 * <p>If no factory method names are declared, a method within the test class
 	 * that has the same name as the test method will be used as the factory
-	 * method by default.
+	 * method by default in case this annotation is applied to a
+	 * {@code @ParameterizedTest} method. For {@code @ParameterizedContainer}
+	 * classes, at least one method name must be declared explicitly.
 	 *
 	 * <p>For further information, see the {@linkplain MethodSource class-level Javadoc}.
 	 */
