+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/provider/NullAndEmptySource.java
@@ -25,11 +25,13 @@ import org.apiguardian.api.API;
  * the functionality of {@link NullSource @NullSource} and
  * {@link EmptySource @EmptySource}.
  *
- * <p>Annotating a {@code @ParameterizedTest} method with
- * {@code @NullAndEmptySource} is equivalent to annotating the method with
- * {@code @NullSource} and {@code @EmptySource}.
+ * <p>Annotating a {@code @ParameterizedContainer} class or
+ * {@code @ParameterizedTest} method with {@code @NullAndEmptySource} is
+ * equivalent to annotating the method with both {@code @NullSource} and
+ * {@code @EmptySource}.
  *
  * @since 5.4
+ * @see org.junit.jupiter.params.ParameterizedContainer
  * @see org.junit.jupiter.params.ParameterizedTest
  * @see NullSource
  * @see EmptySource
