+++ jupiter-tests/src/test/java/org/junit/jupiter/params/provider/AnnotationBasedArgumentsProviderTests.java
@@ -24,14 +24,16 @@ import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.params.support.ParameterDeclarations;
 
 @DisplayName("AnnotationBasedArgumentsProvider")
 class AnnotationBasedArgumentsProviderTests {
 
 	private final AnnotationBasedArgumentsProvider<CsvSource> annotationBasedArgumentsProvider = new AnnotationBasedArgumentsProvider<>() {
 		@Override
-		protected Stream<? extends Arguments> provideArguments(ParameterDeclarations parameters,
-				ExtensionContext context, CsvSource annotation) {
+		protected Stream<? extends Arguments> provideArguments(
+				org.junit.jupiter.params.support.ParameterDeclarations parameters, ExtensionContext context,
+				CsvSource annotation) {
 			return Stream.of(Arguments.of(annotation));
 		}
 	};
@@ -47,7 +49,7 @@ class AnnotationBasedArgumentsProviderTests {
 	@DisplayName("should invoke the provideArguments template method with the accepted annotation")
 	void shouldInvokeTemplateMethodWithTheAnnotationProvidedToAccept() {
 		var spiedProvider = spy(annotationBasedArgumentsProvider);
-		var parameters = mock(ParameterDeclarations.class);
+		var parameters = mock(org.junit.jupiter.params.support.ParameterDeclarations.class);
 		var extensionContext = mock(ExtensionContext.class);
 		var annotation = csvSource("0", "1", "2");
